Move all response related to `response.py`
All response related from http_api_helpers.py
was moved to `response.py`
