+++ server/app/interfaces/discovery.py
@@ -6,15 +6,14 @@ from basyx.aas import model
 
 from .. import server_model
 from ..adapter.jsonization import ServerAASToJsonEncoder
-from ..http_api_helpers import APIResponse, http_exception_to_response, get_response_type, HTTPApiDecoder
 
 from werkzeug.routing import MapAdapter, Rule, Submount
-from ..http_api_helpers import Base64URLConverter, APIResponse, XmlResponse, JsonResponse, XmlResponseAlt, Message, MessageType, Result, HTTPApiDecoder, get_response_type, http_exception_to_response, is_stripped_request
-from typing import Callable, Dict, Iterable, Iterator, List, Optional, Type, TypeVar, Union, Tuple, Set
+from ..http_api_helpers import Base64URLConverter, HTTPApiDecoder
+from ..response import get_response_type, http_exception_to_response
+from typing import Dict, Iterable, List, TypeVar, Set
 
 import abc
 
-import copy
 from pymongo import MongoClient
 from pymongo.collection import Collection
